requirements.txt
alembic==1.8.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==0.12.5
Flask-APIBlueprint==1.0.0
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
hurry.filesize==0.9
idna==2.10
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.2.1
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
mysql-connector-python==8.0.24
mysqlclient==2.1.0
packaging==21.3
protobuf==3.16.0
psutil==5.9.2
PyJWT==2.1.0
pyparsing==3.0.9
python-dotenv==0.17.1
python-logstash==0.4.6
redis==3.5.3
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.4
Werkzeug==0.16.1
zipp==3.8.1
